Move lifecycle query to own file
