+++ back/src/forms/resolvers.ts
@@ -1,10 +1,5 @@
 import { getUserCompanies } from "../companies/queries";
-import {
-  prisma,
-  StatusLogConnection,
-  Company,
-  Status,
-} from "../generated/prisma-client";
+import { prisma, Status } from "../generated/prisma-client";
 import { unflattenObjectFromDb } from "./form-converter";
 import {
   markAsProcessed,
@@ -22,11 +17,7 @@ import { updateTransporterFields } from "./mutations/updateTransporterFields";
 import { formPdf } from "./queries/form-pdf";
 import forms from "./queries/forms";
 import { formsRegister } from "./queries/forms-register";
-import {
-  ForbiddenError,
-  UserInputError,
-  AuthenticationError,
-} from "apollo-server-express";
+import { AuthenticationError } from "apollo-server-express";
 import { stateSummary } from "./queries/state-summary";
 import {
   QueryResolvers,
@@ -38,47 +29,7 @@ import {
 } from "../generated/graphql/types";
 import { transportSegments } from "./queries/segments";
 
-// formsLifeCycle fragment
-const statusLogFragment = `
-  fragment StatusLogPaginated on StatusLogConnection {
-    aggregate {
-      count
-    }
-    pageInfo {
-      hasNextPage
-      hasPreviousPage
-      startCursor
-      endCursor
-    }
-    edges {
-      node {
-        id
-        loggedAt
-        status
-        updatedFields
-        form {
-          id
-          readableId
-        }
-        user {
-          id
-          email
-        }
-      }
-    }
-  }
-`;
-
-const companyFragment = `
-fragment Company on CompanyAssociation {
-  company {
-    id
-    siret
-  }
-}
-`;
-
-const PAGINATE_BY = 100;
+import { formsLifecycle } from "./queries/formsLifecycle";
 
 const queryResolvers: QueryResolvers = {
   form: async (_, { id }) => {
@@ -91,81 +42,9 @@ const queryResolvers: QueryResolvers = {
     return unflattenObjectFromDb(dbForm);
   },
   forms: (_parent, args, context) => forms(context.user.id, args),
-  formsLifeCycle: async (
-    _parent,
-    { siret, loggedAfter, loggedBefore, cursorAfter, cursorBefore, formId },
-    context
-  ) => {
-    const userId = context.user.id;
 
-    const userCompanies = await prisma
-      .companyAssociations({ where: { user: { id: userId } } })
-      .$fragment<{ company: Pick<Company, "id" | "siret"> }[]>(companyFragment)
-      .then((associations) => associations.map((a) => a.company));
+  formsLifeCycle: formsLifecycle,
 
-    // User must be associated with a company
-    if (!userCompanies.length) {
-      throw new ForbiddenError(
-        "Vous n'êtes pas autorisé à consulter le cycle de vie des bordereaux."
-      );
-    }
-    // If user is associated with several companies, siret is mandatory
-    if (userCompanies.length > 1 && !siret) {
-      throw new UserInputError(
-        "Vous devez préciser pour quel siret vous souhaitez consulter",
-        {
-          invalidArgs: ["siret"],
-        }
-      );
-    }
-    // If requested siret does not belong to user, raise an error
-    if (!!siret && !userCompanies.map((c) => c.siret).includes(siret)) {
-      throw new ForbiddenError(
-        "Vous n'avez pas le droit d'accéder au siret précisé"
-      );
-    }
-    // Select user company matching siret or get the first
-    const selectedCompany =
-      userCompanies.find((uc) => uc.siret === siret) || userCompanies.shift();
-
-    const SEALED: Status = "SEALED";
-
-    const formsFilter = {
-      OR: [
-        { owner: { id: userId } },
-        { recipientCompanySiret: selectedCompany.siret },
-        { emitterCompanySiret: selectedCompany.siret },
-        {
-          transporterCompanySiret: selectedCompany.siret,
-          status: SEALED,
-        },
-      ],
-    };
-    const statusLogsCx = await prisma
-      .statusLogsConnection({
-        orderBy: "loggedAt_DESC",
-        first: PAGINATE_BY,
-        after: cursorAfter,
-        before: cursorBefore,
-        where: {
-          loggedAt_not: null,
-          loggedAt_gte: loggedAfter,
-          loggedAt_lte: loggedBefore,
-          form: { ...formsFilter, isDeleted: false, id: formId },
-        },
-      })
-      .$fragment<
-        StatusLogConnection & {
-          aggregate: { count: number };
-        }
-      >(statusLogFragment);
-
-    return {
-      statusLogs: statusLogsCx.edges.map((el) => el.node),
-      ...statusLogsCx.pageInfo,
-      count: statusLogsCx.aggregate.count,
-    };
-  },
   stats: async (_parent, _args, context) => {
     const userId = context.user.id;
     const userCompanies = await getUserCompanies(userId);
